refactor: queue 처리에 대한 메타데이터 추가
