+++ charmander-infra/charmander-file-strorage/src/main/java/org/cm/infra/mediaconvert/queue/impl/AwsWavCombineQueue.java
@@ -1,38 +1,31 @@
-package org.cm.infra.mediaconvert.queue;
+package org.cm.infra.mediaconvert.queue.impl;
 
-import java.util.HashMap;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.cm.infra.mediaconvert.queue.WavVideoCombineQueue.VideoSource;
+import org.cm.infra.mediaconvert.queue.WavCombineQueue;
 import org.cm.infra.property.MediaConvertProperty;
 import org.cm.infra.property.S3URLProperty;
 import org.cm.infra.utils.MetadataConverter;
 import org.springframework.stereotype.Component;
 import software.amazon.awssdk.services.mediaconvert.MediaConvertClient;
 import software.amazon.awssdk.services.mediaconvert.model.AacCodingMode;
+import software.amazon.awssdk.services.mediaconvert.model.AacRateControlMode;
 import software.amazon.awssdk.services.mediaconvert.model.AacSettings;
+import software.amazon.awssdk.services.mediaconvert.model.AacSpecification;
+import software.amazon.awssdk.services.mediaconvert.model.AacVbrQuality;
 import software.amazon.awssdk.services.mediaconvert.model.AccelerationMode;
 import software.amazon.awssdk.services.mediaconvert.model.AccelerationSettings;
 import software.amazon.awssdk.services.mediaconvert.model.AudioCodec;
 import software.amazon.awssdk.services.mediaconvert.model.AudioCodecSettings;
-import software.amazon.awssdk.services.mediaconvert.model.AudioDefaultSelection;
 import software.amazon.awssdk.services.mediaconvert.model.AudioDescription;
-import software.amazon.awssdk.services.mediaconvert.model.AudioSelector;
 import software.amazon.awssdk.services.mediaconvert.model.BillingTagsSource;
 import software.amazon.awssdk.services.mediaconvert.model.ContainerSettings;
 import software.amazon.awssdk.services.mediaconvert.model.ContainerType;
 import software.amazon.awssdk.services.mediaconvert.model.CreateJobRequest;
+import software.amazon.awssdk.services.mediaconvert.model.CreateJobResponse;
 import software.amazon.awssdk.services.mediaconvert.model.DestinationSettings;
 import software.amazon.awssdk.services.mediaconvert.model.FileGroupSettings;
-import software.amazon.awssdk.services.mediaconvert.model.H264RateControlMode;
-import software.amazon.awssdk.services.mediaconvert.model.H264SceneChangeDetect;
-import software.amazon.awssdk.services.mediaconvert.model.H264Settings;
-import software.amazon.awssdk.services.mediaconvert.model.Input;
-import software.amazon.awssdk.services.mediaconvert.model.InputTimecodeSource;
 import software.amazon.awssdk.services.mediaconvert.model.JobSettings;
-import software.amazon.awssdk.services.mediaconvert.model.MotionImageInserter;
-import software.amazon.awssdk.services.mediaconvert.model.MotionImageInsertionMode;
-import software.amazon.awssdk.services.mediaconvert.model.MotionImageInsertionOffset;
-import software.amazon.awssdk.services.mediaconvert.model.MotionImagePlayback;
 import software.amazon.awssdk.services.mediaconvert.model.Mp4Settings;
 import software.amazon.awssdk.services.mediaconvert.model.Output;
 import software.amazon.awssdk.services.mediaconvert.model.OutputGroup;
@@ -40,32 +33,25 @@ import software.amazon.awssdk.services.mediaconvert.model.OutputGroupSettings;
 import software.amazon.awssdk.services.mediaconvert.model.OutputGroupType;
 import software.amazon.awssdk.services.mediaconvert.model.S3DestinationSettings;
 import software.amazon.awssdk.services.mediaconvert.model.S3StorageClass;
-import software.amazon.awssdk.services.mediaconvert.model.ScalingBehavior;
 import software.amazon.awssdk.services.mediaconvert.model.StatusUpdateInterval;
 import software.amazon.awssdk.services.mediaconvert.model.TimecodeConfig;
 import software.amazon.awssdk.services.mediaconvert.model.TimecodeSource;
-import software.amazon.awssdk.services.mediaconvert.model.VideoCodec;
-import software.amazon.awssdk.services.mediaconvert.model.VideoCodecSettings;
-import software.amazon.awssdk.services.mediaconvert.model.VideoDescription;
-
 
 @Component
 @RequiredArgsConstructor
-public class VideoOverlayCombineQueue {
+public class AwsWavCombineQueue implements WavCombineQueue {
 
     private final MediaConvertClient mediaConvertClient;
     private final S3URLProperty s3UrlProperty;
     private final MediaConvertProperty mediaConvertProperty;
 
     public <T> String offer(
-            VideoSource videoSource,
-            VideoSource overlaySource,
             String fileId,
+            List<WavCombineQueue.AudioSource> audioSources,
             T input
     ) {
         CreateJobRequest jobRequest = CreateJobRequest.builder()
-                // TODO 적절한 큐로 바꿀 것
-                .queue(mediaConvertProperty.queue().sceneCombine())
+                .queue(mediaConvertProperty.queue().voiceCombine())
                 .role(mediaConvertProperty.userArn())
                 .userMetadata(MetadataConverter.convert(input))
                 .settings(JobSettings.builder()
@@ -77,8 +63,7 @@ public class VideoOverlayCombineQueue {
                                 .outputGroupSettings(OutputGroupSettings.builder()
                                         .type(OutputGroupType.FILE_GROUP_SETTINGS)
                                         .fileGroupSettings(FileGroupSettings.builder()
-                                                .destination(
-                                                        s3UrlProperty.videoContentsLocator().combineLocation(fileId))
+                                                .destination(s3UrlProperty.ttsContentsLocator().combineLocation(fileId))
                                                 .destinationSettings(DestinationSettings.builder()
                                                         .s3Settings(S3DestinationSettings.builder()
                                                                 .storageClass(S3StorageClass.STANDARD)
@@ -91,65 +76,36 @@ public class VideoOverlayCombineQueue {
                                                 .container(ContainerType.MP4)
                                                 .mp4Settings(Mp4Settings.builder().build())
                                                 .build())
-                                        .videoDescription(VideoDescription.builder()
-                                                .width(1280)
-                                                .height(720)
-                                                .scalingBehavior(ScalingBehavior.FIT)
-                                                .sharpness(100)
-                                                .codecSettings(VideoCodecSettings.builder()
-                                                        .codec(VideoCodec.H_264)
-                                                        .h264Settings(H264Settings.builder()
-                                                                .maxBitrate(6000000)
-                                                                .rateControlMode(H264RateControlMode.QVBR)
-                                                                .sceneChangeDetect(
-                                                                        H264SceneChangeDetect.TRANSITION_DETECTION)
-                                                                .build())
-                                                        .build())
-                                                .build())
                                         .audioDescriptions(AudioDescription.builder()
                                                 .audioSourceName("Audio Selector 1")
                                                 .codecSettings(AudioCodecSettings.builder()
                                                         .codec(AudioCodec.AAC)
                                                         .aacSettings(AacSettings.builder()
-                                                                .bitrate(96000)
+                                                                .vbrQuality(AacVbrQuality.HIGH)
+                                                                .rateControlMode(AacRateControlMode.VBR)
                                                                 .codingMode(AacCodingMode.CODING_MODE_2_0)
                                                                 .sampleRate(48000)
+                                                                .specification(AacSpecification.MPEG4)
                                                                 .build())
                                                         .build())
                                                 .build())
-                                        .extension("mp4")
+                                        .extension("wav")
+                                        .nameModifier("output")
                                         .build())
                                 .build())
-                        .inputs(Input.builder()
-                                .audioSelectors(new HashMap<>() {{
-                                    put("Audio Selector 1", AudioSelector.builder()
-                                            .defaultSelection(AudioDefaultSelection.DEFAULT)
-                                            .build());
-                                }})
-                                .timecodeSource(InputTimecodeSource.ZEROBASED)
-                                .fileInput(videoSource.s3Location())
-                                .build())
-                        .motionImageInserter(MotionImageInserter.builder()
-                                .insertionMode(MotionImageInsertionMode.MOV)
-                                .input(overlaySource.s3Location())
-                                .offset(MotionImageInsertionOffset.builder()
-                                        .imageX(overlaySource.x())
-                                        .imageY(overlaySource.y())
-                                        .build())
-                                .playback(MotionImagePlayback.REPEAT)
-                                .build())
-                        .followSource(1)
+                        .inputs(audioSources.stream().map(WavCombineQueue.AudioSource::toInput).flatMap(List::stream)
+                                .toList())
                         .build())
-                .billingTagsSource(BillingTagsSource.JOB)
                 .accelerationSettings(AccelerationSettings.builder()
                         .mode(AccelerationMode.DISABLED)
                         .build())
-                .statusUpdateInterval(StatusUpdateInterval.SECONDS_10)
+                .statusUpdateInterval(StatusUpdateInterval.SECONDS_60)
                 .priority(0)
+                .billingTagsSource(BillingTagsSource.JOB)
                 .build();
 
-        return mediaConvertClient.createJob(jobRequest)
-                .job()
-                .id();
+        CreateJobResponse response = mediaConvertClient.createJob(jobRequest);
+
+        return response.job().id();
     }
 }
